Update cmake-js with fork
Building iohook for Electron v4 is going to require some updates to `cmake-js`, some of which haven't been merged yet. This PR change the branch to one that merged those PRs.

Note:: DO NOT MERGE THIS PR IN IT'S CURRENT FORM.

If we're going with a fork we should have it be based off wilix-team's repo instead of mine.